+++ packages/analytics_hub/lib/src/core/interception/context/context_entry.dart
@@ -1,4 +1,4 @@
-/// Marker  class for typed context entries.
+/// Marker class for typed context entries.
 abstract class ContextEntry {
   /// Creates a typed context entry marker.
   const ContextEntry();